third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-2.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

async function test() {
  var test = SourcesTestRunner.testJavascriptOutline;

  TestRunner.markStep('testNamedFunctionVariableAssignment');
  await test('var a = function foo(...bar) { }');

  TestRunner.markStep('testArrowFunction');
  await test('var a = x => x + 2');

  TestRunner.markStep('testArrowFunctionWithMultipleArguments');
  await test('var a = (x, y) => x + y');

  TestRunner.markStep('testInnerFunctions');
  await test('function foo(){ function bar() {} function baz() { }}');

  TestRunner.markStep('testObjectProperties');
  await test('x = { run: function() { }, get count() { }, set count(value) { }}');

  TestRunner.completeTest();
}

</script>

</head>

<body onload="runTest()">
<p>Verify javascript outline</p>
</body>
</html>
