third_party/WebKit/LayoutTests/http/tests/devtools/sources/outline-javascript-2-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verify javascript outline


Running: testNamedFunctionVariableAssignment
Text:
    var a = function foo(...bar) { }
Outline:
    a(...bar):0:4

Running: testArrowFunction
Text:
    var a = x => x + 2
Outline:
    a(x):0:4

Running: testArrowFunctionWithMultipleArguments
Text:
    var a = (x, y) => x + y
Outline:
    a(x, y):0:4

Running: testInnerFunctions
Text:
    function foo(){ function bar() {} function baz() { }}
Outline:
    foo():0:9
    bar():0:25
    baz():0:43

Running: testObjectProperties
Text:
    x = { run: function() { }, get count() { }, set count(value) { }}
Outline:
    run():0:6
    get count():0:31
    set count(value):0:48
